api-client cleanup
