+++ apps/fishing-map/utils/async-slice.ts
@@ -6,6 +6,7 @@ import {
   createEntityAdapter,
   IdSelector,
 } from '@reduxjs/toolkit'
+import { ResponseError } from '@globalfishingwatch/api-client'
 
 export enum AsyncReducerStatus {
   Idle = 'idle',
@@ -19,9 +20,7 @@ export enum AsyncReducerStatus {
   Error = 'error',
 }
 
-export type AsyncError<Metadata = Record<string, any>> = {
-  status?: number // HHTP error codes
-  message?: string
+export type AsyncError<Metadata = Record<string, any>> = ResponseError & {
   metadata?: Metadata
 }
 
@@ -38,7 +37,7 @@ export type AsyncReducer<T = any> = {
 export const asyncInitialState: AsyncReducer = {
   status: AsyncReducerStatus.Idle,
   statusId: null,
-  error: {},
+  error: {} as AsyncError,
   ids: [],
   currentRequestIds: [],
   entities: {},
@@ -55,7 +54,7 @@ const getRequestIdsOnFinish = (currentRequestIds: string[], action: any) => {
 export const createAsyncSlice = <
   T,
   U extends { id: AsyncReducerId },
-  Reducers extends SliceCaseReducers<T> = SliceCaseReducers<T>,
+  Reducers extends SliceCaseReducers<T> = SliceCaseReducers<T>
 >({
   name = '',
   initialState = {} as T,
