+++ libs/api-client/src/index.ts
@@ -2,4 +2,5 @@ export * from './utils/url'
 export * from './utils/search'
 export * from './utils/errors'
 export * from './utils/thinning'
+export * from './config'
 export * from './api-client'
